+++ addon/utils/normalize-object.ts
@@ -1,4 +1,4 @@
-import isObject from '../utils/is-object';
+import isObject from './is-object';
 
 /**
  * traverse through target and unset `value` from leaf key so can access normally
